Add files via upload
This is my first very simple project. As simple as it may be, I was very proud of it due to the fact that it was the first time I ever coded anything without any sort of references and only on the basis of my highly limited knowledge thus far.